Feat: Add Resources method to Builder
Feature:
* Add IEmbeddedResourceBuilder#Resources()
* Add IEmbeddedResourceItemBuilder#Resources()

This allows you to pass an Enumerable<IBuilder> to add to the
embedded resource.

As of this commit, the method will crash if the passed
Enumerable is empty. This is because the .Resource()
method returns IEmbeddedResourceItemBuilder
instead of IEmbeddedResourceBuilder. That prevents
exiting early if the passed Enumerable is empty because
the current builder's type does not match.

closes #31