Feat: Add .Resources() to Embedded Builder (#32)
* Feat: Add Resources method to Builder

Feature:
* Add IEmbeddedResourceBuilder#Resources()
* Add IEmbeddedResourceItemBuilder#Resources()

This allows you to pass an Enumerable<IBuilder> to add to the
embedded resource.

As of this commit, the method will crash if the passed
Enumerable is empty. This is because the .Resource()
method returns IEmbeddedResourceItemBuilder
instead of IEmbeddedResourceBuilder. That prevents
exiting early if the passed Enumerable is empty because
the current builder's type does not match.

closes #31

* refactor: allow passing null to EmbeddedResourceItemBuilder

This refactor preserves exposed methods to
prevent passing null intentionally.
Instead, this is done internally in case
an enumerable list of builders is empty.

When building, embedded resources that are null
are skipped.